presentations/test-the-web-forward-tutorial/index.html

<!doctype HTML>  
<html lang="en">
	
	<head>
		<meta charset="utf-8">
		
		<title>Test the Web Forward</title>
        <meta name="apple-mobile-web-app-capable" content="yes" />
		<meta name="apple-mobile-web-app-status-bar-style" content="black-translucent" />
		
		<link href='http://fonts.googleapis.com/css?family=Lato:400,700,400italic,700italic' rel='stylesheet' type='text/css'>
		
		<link rel="stylesheet" href="css/reset.css">
		<link rel="stylesheet" href="css/main.css">
		<link rel="stylesheet" href="css/print.css" type="text/css" media="print">

		<link rel="stylesheet" href="lib/textmate.css">
		<link rel="stylesheet" href="css/adobe.css">
	</head>
	
	<body>
		
		<div class="reveal">

			<!-- Used to fade in a background when a specific slide state is reached -->
			<div class="state-background"></div>
			
			<!-- Any section element inside of this container is displayed as a slide -->
			<div class="slides">
				<section>
					<h2>W3C Test Writing Tutorials</h2>
					<img src="assets/images/W3C_logo.jpg">
					<script>
						// Delicously hacky. Look away.
						if( navigator.userAgent.match( /(iPhone|iPad|iPod|Android)/i ) ) document.write( '<p style="color: rgba(0,0,0,0.3); text-shadow: none;">('+'Tap to navigate'+')</p>' );
					</script>
				</section>
				
				 <section data-state="screenshotPage">
                    <h2>Setup</h2>
                    <p class="font2">Complete all steps before proceeding</p>
                    <section>
                        <br>
                        <ul style="line-height:50px">
                            <li>Step 1: Create a CSSWG Test Manager Account<br><a href="https://test.csswg.org/shepherd/register/" style="font-size:16px">https://test.csswg.org/shepherd/register/</a> 
                                <ul>
                                    <li>Select the check box next to "Request Repository Write Access"</li>
                                </ul>
                            </li>
                            <li>Step 2: Install Mercurial<br><a href="http://wiki.csswg.org/tools/hg/install" style="font-size:16px">http://wiki.csswg.org/tools/hg/install</a></li>
                            <li>Step 3: Setup Mercurial Preferences<br><a href="http://wiki.csswg.org/tools/hg#setting-up-mercurial-preferences" style="font-size:16px">http://wiki.csswg.org/tools/hg#setting-up-mercurial-preferences</a></li>   
                        </ul>
                        <br><br>
                        <p class="font5">Press down for additional steps</p>
                    </section>
                    <section>
                        <br>
                        <ul style="line-height:50px"> 
                            <li>Step 4: Clone the CSSWG Test Suite Repository<br><a href="http://wiki.csswg.org/tools/hg#initial-clone-of-the-central-repository" style="font-size:16px">http://wiki.csswg.org/tools/hg#initial-clone-of-the-central-repository</a></li>
                            <li>Step 5: Clone the Resources directory<br><a href="http://wiki.csswg.org/tools/hg#initial-clone-of-the-central-repository" style="font-size:16px">http://wiki.csswg.org/tools/hg#initial-clone-of-the-central-repository</a></li>
                        </ul>                        
                    </section>
                </section>
				
				<section data-state="screenshotPage">
				    <br>
					<h2>Reftest Overview</h2>
					<blockquote>
					<ul>
					    <li>Two files: Test file + Reference file</li>
					    <li>Test file uses the feature you're testing</li>
					    <li>Reference file is an exact visual match without using the feature you're testing</li>
						<li>Self-describing (works both manually and with automation)</li>
						<li> Cross-browser &amp; cross-platform</li>
					</ul>
					</blockquote>
				</section>
				
				<section data-state="screenshotPage">
				    <br>
    				<h2>Directory structure</h2>
    				<section>
    				    <br>
    				    
    				    <ul>
    				    <li style="list-style-type: none">CSSWG Test Repository</li>
    				    <ul>
    				        <li>test</li>
    				        <ul>
    				            <li>contributors</li>
						        <ul>
						            <li>ttwf</li>
						            <ul>
						                <li><tt><em>your_csswg_id</em></tt></li>
						                <ul>
						                    <li>incoming</li>
    						                <li>submitted</li>
						                </ul>
						            </ul>
						        </ul>
						    </ul>
						</ul>
						</ul>
					</section>
					<section>
					    <img src="assets/images/directoryStructure.png"/>
					</section>
    			</section>
    			<section data-state="screenshotPage">
    			    <br>
    				<h2>Sample Test Case</h2>
    				<section>
    					<br><br>
    					<p style="font-size: 28pt">Basic test for the transform property described in the CSS3 Transforms spec using the <code>translate()</code> function</p>
    					<br>
    					<br>
    					<p>Spec Links:</p>
    					<p style="font-size: 18pt">
    					    <a href="http://dev.w3.org/csswg/css3-transforms/#transform-property" target="top">http://dev.w3.org/csswg/css3-transforms/#transform-property</a>
    					    <br>
    					    <a href="http://dev.w3.org/csswg/css3-transforms/#two-d-transform-functions" target="top">http://dev.w3.org/csswg/css3-transforms/#two-d-transform-functions</a>
					    </p>
					</section>
					<section>
					    <img src="assets/images/specAssertion.png">
					</section>
	            </section>
				<section data-state="screenshotPage">
                      <section>
                          <br>
                          <h2>Step 1 - OSX</h2>
                          <p class="font5">Press down-arrow for Windows instructions</p>
              	        <br>
              		    <p class="font2">Update your local CSSWG test repository</p>
              		    <p style="font-size:22px">Complete this step unless you just cloned the repository during setup</p>
              		    <br>

              	        <p style="font-size:30px">From a terminal window, type the following:</p>
              		    <div class="console">
              		        <div class="consoleText">
              			        <br><br>
              		            <p>$ cd <span class="localPath">[path to repo]</span>/test</p>
              		            <p>$ hg pull -u</p>
              		            <p>$ ls -lR contributors/ttwf/samples</p>
              		        </div>
              	        </div>
              	        <br><br><br>
              	        <div class="bottomNote">
              	            <p class="font5">
              	                Note: These files can also be downloaded from here:
              	                <br>
              	                <a href="assets/resources/ttwf-reftest-tutorial-starter.zip">ttwf-reftest-tutorial-starter.zip</a>
              	            </p>
              	        </div>
                      </section>  
                      
                      <section>
                          <br>
                          <h2>Step 1 - Windows</h2>
                          <p class="font5">Press up-arrow for OSX instructions</p>
              	        <br>
              		    <p class="font2">Update your local CSSWG test repository</p>
              		    <p style="font-size:22px">Complete this step unless you just cloned the repository during setup</p>
              		    <br>

              	        <p style="font-size:30px">From a DOS command window, type the following:</p>
              		    <div class="console">
              		        <div class="dosText">
              			        <br><br>
              		            <p>> cd <span class="localPath">[path to repo]</span>\test</p>
              		            <p>> hg pull -u</p>
              		            <p>> dir contributors\ttwf\samples</p>
              		        </div>
              	        </div>
              	        <br><br><br>
              	        <div class="bottomNote">
              	            <p class="font5">
              	                Note: These files can also be downloaded from here:
              	                <br>
              	                <a href="assets/resources/ttwf-reftest-tutorial-starter.zip">ttwf-reftest-tutorial-starter.zip</a>
              	            </p>
              	        </div>                                                 
              	    </section>
              	    
                  </section>

                  <section data-state="screenshotPage">
                      <section>
                          <br>
              	         <h2>Step 2 - OSX</h2>
              	         <p class="font5">Press down-arrow for Windows instructions</p>
              		     <br>
              		     <p class="font2">Create a directory for your new test</p>
              		     <br>
              			 <div class="console">
              			     <div class="consoleText">
                				     <br><br><br>
                				     <p>$ cd <span class="localPath">[path to repo]</span>/test/contributors/ttwf/</p>
                				     <p>$ mkdir -p <span class="localPath">yourName</span>/incoming</p>
                				     <p>$ cd <span class="localPath">yourName</span>/incoming</p>
                			     </div>
              		     </div>
              	    </section>
              	    <section>
              	        <br>
              	        <h2>Step 2 - Windows</h2>
              	        <p class="font5">Press up-arrow for OSX instructions</p>
              		    <br>
              		    <p class="font2">Create a directory for your new test</p>
              		    <br>
              			<div class="console">
              			    <div class="dosText">
                 				     <br>
                 				     <br>
                 				     <br>
                 				     <p>> cd <span class="localPath">[path to repo]</span>\test\contributors\ttwf</p>
                 				     <p>> mkdir <span class="localPath">yourName</span>\incoming</p>
                 				     <p>> cd <span class="localPath">yourName</span>\incoming</p>
                 			    </div>
              		    </div>
              	    </section>
                  </section>

                  <section data-state="screenshotPage">
                      <section>
                          <br>
              	         <h2>Step 3 - OSX</h2>
              	         <p class="font5">Press down-arrow for Windows instructions</p>
              		     <br>
              		     <p class="font2">Copy the starter files to your incoming directory</p>
              		     <br>
              			 <div class="console">
              			     <div class="consoleText">
                				     <br>
                				     <br>
                				     <p><span class="localPath">From your incoming directory created in previous step</span></p>
                			         <p>$ cp -R ../../samples/* .</p>
                			     </div>
              		     </div>
                      </section>
                      <section>
                          <br>
                          <h2>Step 3 - Windows</h2>
              	         <p class="font5">Press up-arrow for OSX instructions</p>
              		     <br>
              		     <p class="font2">Copy the starter files to your incoming directory</p>
              		     <br>
              			 <div class="console">
              			     <div class="dosText">
                				     <br>
                				     <br>
                				     <p><span class="localPath">From your incoming directory created in previous step</span></p>
                			         <p>> xcopy /e ..\..\samples .</p>
                			     </div>
              		     </div>                        
                      </section>
                  </section>

                  <section data-state="screenshotPage">
                     <br>
                     <h2>Step 4</h2>
              	     <br>
                     <p class="font2">Open the ttwf-reftest-starter-001.html file in a editor</p>
                  </section>

                  <section data-state="screenshotPage">
                     <br>
              	     <h2>Step 5.1 </h2>
              	     <br>
              	     <p class="font2">Add the test metadata</p>
              	     <p class="font3">Title & Author</p>
              	     				    <pre>
                     					<code class="html">
                     <span class="expand">
    &lt;!DOCTYPE html&gt;
     &lt;html&gt;
         &lt;head&gt;
         &lt;title&gt;CSS Transforms Test: transform property with translate function&lt;/title&gt;
         &lt;link rel="author" title="Your Name" href="mailto:youremail@address.com"&gt;
             . . .			
                         </span>		        
                     				    </code>
                     				    </pre>
              	     
              		<div class="bottomNote3">
                          <p class="font5">
                              Note: Each metadata element is described in detail on the CSSWG Wiki<br>
                              <a href="http://wiki.csswg.org/test/format#template-details">http://wiki.csswg.org/test/format#template-details</a>
                          </p>
                      </div>
                  </section>
                  
                   <section data-state="screenshotPage">
                        <br>
                	     <h2>Step 5.2</h2>
                	     <br>
                	     <p class="font2">Add the test metadata</p>
                	     <p class="font3">Path to reference file</p>
                	     				    <pre>
                       					<code class="html">
                       <span class="expand">
      &lt;!DOCTYPE html&gt;
       &lt;html&gt;
           &lt;head&gt;
              . . .
              &lt;link rel="match" href="reftest/ttwf-reftest-tutorial-starter-ref.html"&gt;
              . . .
                           </span>		        
                       				    </code>
                       				    </pre>

                		<div class="bottomNote3">
                            <p class="font5">
                                Note: Each metadata element is described in detail on the CSSWG Wiki<br>
                                <a href="http://wiki.csswg.org/test/format#template-details">http://wiki.csswg.org/test/format#template-details</a>
                            </p>
                        </div>
                    </section>
                    
                     <section data-state="screenshotPage">
                             <br>
                    	     <h2>Step 5.3</h2>
                    	     <br>
                    	     <p class="font2">Add the test metadata</p>
                    	     <p class="font3">Spec links</p>
                    	     				    <pre>
                           					<code class="html">
                           <span class="expandPageMed">
&lt;!DOCTYPE html&gt;
&lt;html&gt;
    &lt;head&gt;
        . . .
        &lt;link rel="help" href="http://www.w3.org/TR/css3-transforms/#transform-property"&gt;
        &lt;link rel="help" href="http://www.w3.org/TR/css3-transforms/#two-d-transform-functions"&gt;     
        . . .
                               </span>		        
                           				    </code>
                           				    </pre>

                    		<div class="bottomNote3">
                                <p class="font5">
                                    Note: Each metadata element is described in detail on the CSSWG Wiki<br>
                                    <a href="http://wiki.csswg.org/test/format#template-details">http://wiki.csswg.org/test/format#template-details</a>
                                </p>
                            </div>
                        </section>
                        
                        <section data-state="screenshotPage">
                                 <br>
                        	     <h2>Step 6</h2>
                        	     <br>
                        	     <p class="font2">Add the transform property to the test element</p>
                        	     
                        	         <pre>
                 				    <code class="html">
             <span class="expand2">
                    &lt;<span class="styleTag">style type</span>="text/css"&gt;
                         <span class="styleTag">.greenSquare</span> {
                             <span class="cssProperty">position: absolute;
                             top: 0;
                             left: 0;
                             width: 100<span class="px">px</span>;
                             height: 100<span class="px">px</span>;
                             <span class="highlight">transform: translate(100<span class="px">px</span>,100<span class="px">px</span>);</span>
                         }
                         . . .
                     &lt;/<span class="styleTag">style</span>&gt;</span>				        
                 				    </code>
                 				    </pre>

                         </section>

                         <section data-state="screenshotPage">
                                 <br>
                        	     <h2>Step 7</h2>
                        	     <br>
                        	     <p class="font2">Create an element that is visible only upon failure</p>
                        	      <pre>
                 				    <code class="html">
                                    <span class="expand2">
                     &lt;<span class="styleTag">style type</span>="text/css"&gt;
                         . . .
                         <span class="styleTag">.redSquare</span> {
                             <span class="cssProperty">position: absolute;
                             width: 98<span class="px">px</span>;
                             height: 98<span class="px">px</span>;
                             <span class="highlight">top: 101<span class="px">px</span>;</span>
                             <span class="highlight">left: 101<span class="px">px</span>;</span>
                         }
                     &lt;/<span class="styleTag">style</span>&gt;</span>				        
                 				    </code>
                 				    </pre>

                         </section>

                         <section>
                             <section data-state="screenshotPage">
                                    <h2>Step 8.1</h2>
                        	        <br>
                        	        <p class="font2">Add green fill color to denote passing condition</p>
                        	        <pre>
              				            <code class="html">
                <span class="expand2">
                      <span class="styleTag">.greenSquare</span> {
                          <span class="cssProperty">position: absolute;
                          top: 0;
                          left: 0;
                          width: 100<span class="px">px</span>;
                          height: 100<span class="px">px</span>;
                          transform: translate(100<span class="px">px</span>,100<span class="px">px</span>);
                          <span class="highlight">background: green;</span>
                      }</span>				        
              				    </code>
              				    </pre>
                        </section>
                        <section data-state="screenshotPage">
                            <h2>Step 8.2</h2>
                    	    <br>
                    	    <p class="font2">Add the red fill color to denote failing condition</p>
                    	    <pre>
          				    <code class="html">
                <span class="expand2">
                   <span class="styleTag">.redSquare</span> {
                       <span class="cssProperty">position: absolute;
                       width: 98<span class="px">px</span>;
                       height: 98<span class="px">px</span>;
                       top: 101<span class="px">px</span>;
                       left: 101<span class="px">px</span>;
                       <span class="highlight">background: red;</span>
                   }
              &lt;/<span class="styleTag">style</span>&gt;</span>				        
          				    </code>
          				    </pre>
                    </section>
                        
                        
                       </section>
                        <section data-state="screenshotPage">
                            <br>
                             <h2>Step 9</h2>
                             <br>
                             <p class="font2">Add a statement to the test page to make this a self-describing test</p>
                              <pre>
             				 <code class="html">
         <span class="expand2">
         &lt;body&gt;
             &lt;p&gt;The test passes if there is a green square and no red.&lt;/p&gt;
             . . .
         &lt;/body&gt;
         </span>    				        
             			    </code>
             			    </pre>

                        </section>

                        <section data-state="screenshotPage">
                            <br>
                             <h2>Step 10</h2>
                             <br>
                             <p class="font2">Open the reftests/ttwf-reftest-tutorial-starter-ref.html file in an editor</p>
    
                        </section>

                        <section data-state="screenshotPage">
                            <br>
                             <h2>Step 11</h2>
                             <br>
                             <p class="font2">Add the reference metadata</p>
                             <p class="font3">Title & Author</p>
                             <pre>
                                 
        <span class="expand">
                         	<code class="html">
     &lt;!DOCTYPE html&gt;
     &lt;html&gt;
         &lt;head&gt;
         &lt;title&gt;CSS Transforms Reference file&lt;/title&gt;
         &lt;link rel="author" title="Your Name" href="mailto:youremail@address.com"&gt;
         . . .
                            </span>
                            </code>
   					    </pre>

                    </section>
                    <section data-state="screenshotPage">
                        <br>
                         <h2>Step 12</h2>
                         <br>
                         <p class="font2">Position and fill the reference element so it matches the test file exactly in a passing condition</p>
                         <pre>
     				    <code class="html">
 <span class="expand2">
                     &lt;<span class="styleTag">style type</span>="text/css"&gt;
                         <span class="styleTag">.greenSquare</span> {
                             <span class="cssProperty">position: absolute;
                             width: 100<span class="px">px</span>;
                             height: 100<span class="px">px</span>;
                             <span class="highlight">top: 100;</span>
                             <span class="highlight">left: 100;</span>
                             <span class="highlight">background: green;</span>
                         }
                     &lt;/<span class="styleTag">style</span>&gt;</span>				        
     				    </code>
     				    </pre>

                    </section>

                    <section data-state="screenshotPage">
                         <br>
                         <h2>Step 13</h2>
                         <br>
                         <p class="font2">Add the self-describing statement here, too</p>
                         <pre>
         				 <code class="html">
     <span class="expand2">
         &lt;body&gt;
             &lt;p&gt;The test passes if there is a green square and no red.&lt;/p&gt;
             . . .
         &lt;/body&gt;
     </span>    				        
         			    </code>
         			    </pre>
                    </section>

                    <section data-state="screenshotPage">
                        <br>
                         <h2>Step 14</h2>
                         <br>
                         <p class="font2">Examine your test and reference files in the browser and confirm they are identical</p>
                         <img src="assets/images/testFile.png">
                         <img src="assets/images/refFile.png">
                    </section>
                    <section>
                         <h3>Failing test example</h3>
                         <br>
                         <img src="assets/images/failingRefTest.png">
                    </section>
                    
                     <section>
                         <h2>Vendor prefixes</h2>
                         
                         <section>
                             <br>
                             <ul>
                                 <li>You may need to add a vendor prefix to the CSS property you're testing</li>
                                 <li>If the prefix is necessary, your test will appear to fail without it</li>
                                 <li>However, do not include vendor prefixes when pushing tests to the W3C</li>
                             </ul>    
                             <br>
                             <p class="font5">Press the down-arrow for a vendor-prefixed version of this tutorial</p>
                        </section>
                        
                        <section>
                            <pre>
                            <code class="html">
                              <span class="expand2">
                        <span class="styleTag">.greenSquare</span> {
                              <span class="cssProperty">position: absolute;
                              top: 0;
                              left: 0;
                              width: 100<span class="px">px</span>;
                              height: 100<span class="px">px</span>;
                              <span class="highlight">-webkit-transform: translate(100<span class="px">px</span>,100<span class="px">px</span>);</span>
                              <span class="highlight">-moz-transform: translate(100<span class="px">px</span>,100<span class="px">px</span>);</span>
                              <span class="highlight">-ms-transform: translate(100<span class="px">px</span>,100<span class="px">px</span>);</span>
                              <span class="highlight">-o-transform: translate(100<span class="px">px</span>,100<span class="px">px</span>);</span>
                              background: green;
                          }</span> 
                            </code>
                            </pre>
                        </section>
                        
                    </section>
                    
                    <section data-state="screenshotPage">
                          <section>
                            <br>
                  	        <h2>Step 15 - OSX</h2>
                  	        <p class="font5">Press the down-arrow for Windows instructions</p>
                  	        <br>
                  		    <p class="font2">Add your new test & reference file to the repository</p>
                  			<div class="console">
                  			    <div class="consoleText">
                    				    <br>
                    				    <p>$ cd <span class="localPath">[path to repo]</span>/test/contributors/ttwf/<span class="localPath">yourName</span>/incoming</p>
                    			        <p>$ hg add ttwf-reftest-tutorial-001.html</p>
                    			        <p>$ hg add reftest</p>
                    			        <p>$ hg commit -m "TTWF reftest tutorial"</p>
                    			        <p>$ hg push</p>
                    			    </div>
                  		    </div>
                               <div class="bottomNote2">
                                   <br><br><br>
                  		            <p class="noteText">Note: If you receive the following error:</p> 
                  		            <p class="errorText">abort: push creates new remote head ...</p>
                  		            <p class="noteText">Your local repository is out of sync with the test repository</p>
                  		            <p class="noteText">Do the following to resolve the issue:</p>
                  		    </div>  
                  	        <br><br><br><br><br><br><br>
                              <div class="console">
                  			    <div class="consoleText">
                    				    <p class="localPath">Only complete these steps if you received the error message above</p>
                    			        <p>$ hg pull -u</p>
                    			        <p>$ hg merge</p>
                    			        <p>$ hg commit -m "Merge"</p>
                    			        <p>$ hg push</p>
                    			    </div>
                  		    </div>          			    
                  	    </section>
                  	    <section>
                  	        <br>
                  	        <h2>Step 15 - Windows</h2>
                  	        <p class="font5">Press the up-arrow for OSX instructions</p>
                  	        <br>
                  		    <p class="font2">Add your new test & reference file to the repository</p>
                  			<div class="console">
                  			    <div class="dosText">
                    				    <br>
                    				    <p>> cd <span class="localPath">[path to repo]</span>\test\contributors\ttwf\<span class="localPath">yourName</span>\incoming</p>
                    			        <p>$ hg add ttwf-reftest-tutorial-001.html</p>
                    			        <p>$ hg add reftest</p>
                    			        <p>> hg commit -m "TTWF reftest tutorial"</p>
                    			        <p>> hg push</p>
                    			    </div>
                  		    </div>
                              <div class="bottomNote2">
                                 <br><br><br>
                  	            <p class="noteText">Note: If you receive the following error:</p> 
                  	            <p class="errorText">abort: push creates new remote head ...</p>
                  	            <p class="noteText">Your local repository is out of sync with the test repository</p>
                  	            <p class="noteText">Do the following to resolve the issue:</p>
                  	        </div>  
                  	        <br><br><br><br><br><br><br>
                              <div class="console">
                  			    <div class="dosText">
                    				    <p class="localPath">Only complete these steps if you received the error message above</p>
                    			        <p>> hg pull -u</p>
                    			        <p>> hg merge</p>
                    			        <p>> hg commit -m "Merge"</p>
                    			        <p>> hg push</p>
                    			    </div>
                  		    </div>      			               			    
                  	    </section>
                      </section>
                      <section data-state="screenshotPage">
                         <br>
                  	     <h2>Step 16</h2>
                  	     <br><br>
                  	     <h3>Ask the spec expert to review your test</h3>
                      </section>
                      <section data-state="screenshotPage">
                          <section>
                             <br>
                  		    <h2>Step 17 - OSX</h2>
                  		    <p class="font5">Press down-arrow for Windows instructions</p>
                  		    <br>
                  		    <p class="font2">Incorporate feedback & move test to the submitted folder</p>
                  		    <br>
                  			<div class="consoleLarge">
                  			    <div class="consoleText">
                  			            <br>
                    				    <p>$ cd <span class="localPath">[path to repo]</span>/test/contributors/ttwf/<span class="localPath">yourName</span></p>
                    				    <p>$ hg pull -u</p>
                    				    <p>$ hg merge</p>
                    				    <p>$ hg commit -m "merging"</p>
                    			        <p>$ hg mv incoming/ttwf-reftest-tutorial-001.html submitted/ttwf-reftest-tutorial-001.html</p>
                    			        <p>$ hg mv incoming/reftest/ttwf-reftest-tutorial-ref.html submitted/reftest/ttwf-reftest-tutorial-ref.html</p>
                    			        <p>$ hg commit -m "moved the TTWF reftest tutorial to the submitted folder"</p>
                    			        <p>$ hg push</p>
                    			    </div>
                  		    </div>
                  	    </section>
                          <section>
                            <br>
                  	        <h2>Step 17 - Windows</h2>
                  		    <p class="font5">Press down-arrow for OSX instructions</p>
                  		    <br>
                  		    <p class="font2">Incorporate feedback & move test to the submitted folder</p>
                  		    <br>
                  			<div class="consoleLarge">
                  			    <div class="dosText">
                  			         <br>
                    				    <p>> cd <span class="localPath">[path to repo]</span>\test\contributors\ttwf\<span class="localPath">yourName</span></p>
                    				    <p>> hg pull -u</p>
                    				    <p>> hg merge</p>
                    				    <p>> hg commit -m "merging"</p>
                    			        <p>> hg mv incoming\ttwf-reftest-tutorial-001.html submitted\ttwf-reftest-tutorial-001.html</p>
                    			        <p>> hg mv incoming\reftest/ttwf-reftest-tutorial-ref.html submitted\reftest\ttwf-reftest-tutorial-ref.html</p>
                    			        <p>> hg commit -m "moved the TTWF reftest tutorial to the submitted folder"</p>
                    			        <p>> hg push</p>
                    			    </div>
                  		    </div>
                          </section>          		        
                      </section>
                      
                      
                      <section>
                          <h1>W3C JavaScript Test Tutorial</h1>
                      </section>

      				<section>
      					<h2>JavaScript Test Overview</h2>
      					<br>
      					<ul>
      					    <li>JavaScript tests programmatically verify functionality</li>
      					    <li>JavaScript tests have some advantages over Ref Tests
      					        <ul>
      					            <li>Robust</li>
      					            <li>Flexible</li>
      					            <li>Performant</li>
      					            <li>Powerful</li>
      					        </ul>
      					    </li>
      					    <li>Ref Tests have some advantages over JavaScript tests
      					        <ul>
      					            <li>Tests verify complete rendering</li>
      					            <li>Easy to understand</li>
      					            <li>Avoids introducing additional technologies (JS)</li>
      					        </ul>
      					    </li>
      					</ul>
      			    </section>

      				<section>
      				    <h2>JavaScript Test Overview (cont'd)</h2>
      				    <br>
                          <ul>
          					<li>JavaScript can be used in conjunction with Ref Tests
          					    <ul>
          				            <li>Test can have both a reference file and a JavaScript section</li>
          				            <li>Ref Test portion supports suites without script engines</li>
          				            <li>JavaScript section adds automation</li>
          				        </ul>
          				    </li>
          				    <li>Self-describing RefTests are the preferred test format</li>
          				    <li>JavaScript tests should used in the following scenarios:
          				        <ul>
          				            <li>Testing a JavaScript API or behavior (ex: CSSOM spec)</li>
          				            <li>Adding automation to Ref Tests</li>
          				        </ul>
          				    </li>
          				    <li>W3C provides a framework (testharness) to simplify and standardize JS test creation</li>
      					</ul>
      				</section>

      				<section data-state="screenshotPage">
      				    <br>
  				        <h3>testharness</h3>
  				        <section>
      				        <p style="font-size:24px">testharness.js and testharnessreport.js</p>
      				        <pre>
      				            <code class="html" style="font-size:24px; line-height:26px">
&lt;script src="/resources/testharness.js"&gt;&lt;/script&gt;
&lt;script src="/resources/testharnessreport.js"&gt;&lt;/script&gt;
                                </code>
                            </pre>    
          				    <ul style="line-height:60px">
          				        <li>JavaScript API for making common test assertions</li>
          				        <li>Documented in the comment section of testharness.js</li>
          				        <li>Should be used by all JavaScript tests</li>
          				    </ul>
          				    <br><br>
          				    <p class="font5">Press down for more</p>
      				    </section>
      				    <section>
      				        <ul>
          				        <li>Path must be to resources directory at root level
          				            <ul>
          				                <li>Locally, resources directory can be located anywhere</li> 
          				                <li>Prior to pushing to CSSWG Test repository, the path must match example above</li>
          				            </ul>
          				        </li>
          				        <li>Element with id="log" must exist in test file
          				            <ul>
          				                <li>Test results table will be added to this element</li>
          				            </ul>
          				        </li>
      				        </ul>
      				    </section>
      				</section>

      				<section data-state="screenshotPage">
      			        <section>
          			        <h2>The test function</h2>
              			    <pre>
              			        <code class="html" style="font-size:24px">
            test(test_function, name, properties)
              			        </code>
              			    </pre>
              			    <ul style="line-height:60px">
              			        <li>Used for synchronous tests
              			            <ul>
              			                <li>async_test() is also defined by testharness for asychronous testing</li>
              			            </ul>
              			        </li>
              			        <li>Documented and implemented in testharness.js</li>
              			        <li>Used to define an individual test case in a file
              			            <ul>
              			                <li>Common for a file to have multiple calls to test()</li>
              			            </ul>
              			        </li>   
              			        <li>testharness will display PASS/FAIL for each call to test()</li>
              			    </ul>
              			    <br><br>
              			    <p class="font5">Press down for more</p>
                          </section>

                          <section>
                              <h2>test() Argument 1: test_function</h2>
              			    <pre>
              			        <code class="html" style="font-size:24px">
            test(<span style="color:red">test_function</span>, name, properties)
              			        </code>
              			    </pre>                        
                              <ul style="line-height:60px">
                                  <li>test_function must be an object, not a function call</li>
                                  <li>test_function should be testharness assert method
                                      <ul>
                                          <li>Can also be user defined if no assert method provides the needed functionality</li>
                                      </ul>
                                  </li>
                                  <li>Result from the assertion will determine PASS/FAIL</li>
                                  <li>Usage example:
                                      <pre style="font-size:18px">
                                          <code>
            test(<span class="highlight">function() {assert_true(true)}</span>, name, properties)
                                          </code>
                                      </pre>  
                                  </li>                            
          			        </ul>                   
                          </section>

                         <section>
                              <h2>test() Argument 2: Name</h2>
                              <pre>
                                  <code class="html" style="font-size:24px">
            test(test_function,<span style="color:red"> name</span>, properties)
                                  </code>
                              </pre>                        
                              <ul style="line-height:60px">
                                  <li>Name is a string that identifies the test</li> 
                                  <li>Descriptive test name that conveys what is being tested</li>
              			        <li>Name will be displayed in the test results table</li>
                                  <li>Usage example:
                                      <pre>
                                          <code class="html" style="font-size:18px">
    test(function() {assert_true(true)}, <span class="highlight">"Assert_true with true"</span>, properties)
                                          </code>
                                      </pre>  
                                  </li>
                              </ul>       			                                   
                          </section>

                          <section>
                              <h2>test() Argument 3: Properties</h2>
                              <pre>
                                  <code class="html" style="font-size:24px">
            test(test_function, name,<span style="color:red"> properties</span>)
                                  </code>
                              </pre>                        
                              <ul style="line-height:60px">
                                  <li>Properties is an object that overrides defaults</li>
                                  <li>Recognized properties include timeout and metadata</li>
                                  <li>Usage example:
                                      <pre>
                                          <code class="html" style="font-size:18px; line-height:16px">
 test(function() {assert_true(true)}, "Assert_true with true", <br><br> <span class="highlight">{timeout:1000,</span> <br><br> <span class="highlight">help: 'http://www.w3.org/TR/spec',</span> <br><br> <span class="highlight">author: ['Bill Gates &lt;bgates@msn.com&gt;', 'Steve Jobs http://apple.com/sjobs'],</span><br><br> <span class="highlight">assert: 'This should always be true.']}</span>)
                                          </code>
                                      </pre>                                
                                  </li>    			                    
          			        </ul>
                          </section>
                      </section>

              	    <section data-state="screenshotPage">
              	        <section>
          			        <h2>Assert methods</h2>
              			    <ul>
              			        <li>testharness.js provides a number of assert methods
              			            <ul>
              			                <li>Full list is provided in the comment section of testharness.js</li>
              			                <li>Typically supplied as the test_function for test()</li>
              			            </ul>
              			        </li>
              			        <li>All assert methods contain an optional "description" argument that is only output if the assertion fails
              			            <ul>
              			                <li>Used to provide additional debug information</li>
              			            </ul>
              			        </li>
              			        <li>All asserts must be located in a test()
              			            <ul>
              			                <li>asserts outside won't be detected correctly by the harness and may cause a file to stop testing</li>
              			            </ul>
              			        </li>
              			    </ul>
              			    <br><br>
              			    <p class="font5">Press down for more</p>
          			    </section>

          			    <section>
          			        <h2>Assert Example</h2>
                              <pre>
                                  <code class="html" style="font-size:24px">
        assert_equals(actual, expected, description)
                                  </code>
                              </pre>
                              <ul style="line-height:60px">
                                  <li>actual: The actual value from functionality being tested</li>
                                  <li>expected: The expected value</li>
                                  <li>description (optional): Output only if the assertion fails</li>
                                  <li>Usage example:
                                      <pre style="font-size:16px">
                                          <code class="html">
 assert_equals(getActualData("myElement"), 100, "Expected value for myElement is 100")
                                          </code>
                                      </pre>                                
                                  </li>                            
                              </ul>    			        
          			    </section>
      				</section>
      				
                    <section data-state="screenshotPage">
                        <h2>Sample Test Case</h2>
                            <section>
              					<br>
              					<p style="font-size: 28pt">Basic test for the transform property described in the CSS3 Transforms spec using the translate() function</p>
              					<p class="font4" style="color:red">Similar to Ref Test example, this time testing API instead of rendering</p>
              					<br>
              					<p>Spec Links:</p>
              					<p style="font-size: 18pt">
              					    <a href="http://dev.w3.org/csswg/css3-transforms/#transform-property" target="top">http://dev.w3.org/csswg/css3-transforms/#transform-property</a>
              					    <br>
              					    <a href="http://dev.w3.org/csswg/css3-transforms/#two-d-transform-functions" target="top">http://dev.w3.org/csswg/css3-transforms/#two-d-transform-functions</a>
          					    </p>
                                <br>
                                <p class="font5">Press down for more</p>
                         </section> 
                         <section>
                              <img height="70%" width="70%" src="assets/images/specAssertion.png">
                         </section>       

                      </section>
                      
                      <section data-state="screenshotPage">
          		        <h2>Step 1</h2>
          		        <br>
          			    <p class="font2">A JavaScript starter file was copied during Setup</p>
          			    <p class="font2">Open the <span style="color:red">ttwf-js-tutorial-001.html</span> file in an editor</p>
                        <pre>
                            <code class="html">
&lt;!DOCTYPE html&gt;
&lt;html&gt;
&lt;head&gt;
    &lt;title&gt;CSS Transforms API Test: transform translate&lt;/title&gt;
    &lt;link rel="author" title="Your Name" href="mailto:youremail@address.com"&gt;
    &lt;link rel="help" href="http://www.w3.org/TR/css3-transforms/#transform-property"&gt;
    &lt;link rel="help" href="http://www.w3.org/TR/css3-transforms/#two-d-transform-functions"&gt;
    &lt;meta name="assert" content="The transform should be translate(100px,100px)"&gt;   
&lt;/head&gt;
&lt;body&gt;
&lt;/body&gt;
&lt;/html&gt;                  
                            </code>
                        </pre>
                      </section>

                      <section data-state="screenshotPage">
          			     <h2>Step 2</h2>
              		     <p class="font2">Update Metadata</p>

      			        <ul style="line-height:60px">
      			            <li>Update the author metadata</li>
      			            <li>Add a line for the flags metadata
      			                <ul>
      			                    <li>Flags document a test's prerequisites</li>
      			                    <li>Full list of flags is available on the CSSWG Wiki <a href="http://wiki.csswg.org/test/format#requirement-flags">Test Format</a> page</li>
      			                </ul>
      			            </li>
      			            <li>Example:</li>
      			            <pre style="width:800px">
      			                <code class="html">
&lt;title&gt;CSS Transforms API Test: transform translate&lt;/title&gt;
&lt;link rel="author" <span class="highlight">title="Jim Tester" href="mailto:jimmyt@yahoo.com"&gt;</span>
&lt;link rel="help" href="http://www.w3.org/TR/css3-transforms/#transform-property"&gt;
&lt;link rel="help" href="http://www.w3.org/TR/css3-transforms/#two-d-transform-functions"&gt;
<span class="highlight">&lt;meta name="flags" content="dom"&gt;</span>
&lt;meta name="assert" content="The transform should be translate(100px,100px)"&gt;  			                    
      			                </code>
      			            </pre>   
        				    </ul> 
                      </section>

                      <section data-state="screenshotPage">
              			<h2>Step 3</h2>
          			    <br>
              		    <p class="font2">Import the W3C JavaScript Test Harness files</p> 
              		    <p class="font3">Add the following lines to the starter file above the style block</p>   
  			            <pre style="width:600px">
			                <code class="html">
...
&lt;script src="/resources/testharness.js"&gt;&lt;/script&gt;
&lt;script src="/resources/testharnessreport.js"&gt;&gt;&lt;/script&gt;
...
                            </code>
			            </pre>
                          <p class="font5">Note: During development, the path to the resources directory should match location on local machine</p>
                          <p class="font5" style="font-weight:bold; color:red">The path to these scripts must match example above before pushing to repo</p>
                      </section>
                      
                      <section data-sate="screenshotPage">
                          <h2>Checkpoint</h2>
                          <br>
                          <p class="font2">The head block of your started file should appear as follows</p>
                          <p class="font3">With your actual name and email address for the author metadata</p>
                          <pre>
                              <code class="html">
&lt;!DOCTYPE html&gt;
&lt;html&gt;
&lt;head&gt;
  &lt;title&gt;CSS Transforms API Test: transform translate&lt;/title&gt;
  &lt;link rel="author" title="Your Name" href="mailto:youremail@address.com"&gt;
  &lt;link rel="help" href="http://www.w3.org/TR/css3-transforms/#transform-property"&gt;
  &lt;link rel="help" href="http://www.w3.org/TR/css3-transforms/#two-d-transform-functions"&gt;
  &lt;meta name="flags" content="dom"&gt;
  &lt;meta name="assert" content="The transform should be translate(100px,100px)"&gt; 
  &lt;script src="/resources/testharness.js"&gt;&lt;/script&gt;
  &lt;script src="/resources/testharnessreport.js"&gt;&lt;/script&gt;    
&lt;/head&gt; 
...                                 
                              </code>
                          </pre>
                      </section>

                      <section data-state="screenshotPage">
                         <h2>Step 4</h2>
          			    <br>
              		    <p class="font2">Add a div to be used as the test element to the body</p>
              		    <br>
              		    <pre style="width:600px">
              		        <code class="html">
...
&lt;body&gt;
    &lt;div id="myDiv"&gt;&lt;/div&gt;
&lt;body&gt;
...
              		        </code>
              		    </pre>
              		    <br>
              		    <p class="font2">The transform will be applied to this element</p>
                      </section>


                      <section data-state="screenshotPage">
              			<h2>Step 5</h2>
          			    <br>
              		    <p class="font2">Add a div with id="log" for test results</p>
              		    <br>
              		    <pre style="width:600px">
              		        <code class="html">
...
&lt;body&gt;
    &lt;div id="myDiv"&gt;&lt;/div&gt;
    <span class="highlight">&lt;div id="log"&gt;&lt;/div&gt;</span>
&lt;body&gt;
...
              		        </code>
              		    </pre>
             			     <br>
             			     <p class="font3">testharness will add an HTML table that contains test results to this element</p>
                      </section>
                      
                      <section data-state="screenshotPage">
                           <h2>Step 6</h2>
                           <br>
                           <p class="font2">Create a script block that will contain the JavaScript</p>
                           <br>
             		       <pre style="width:600px">
             		           <code class="html">
...
&lt;body&gt;
   &lt;div id="myDiv"&gt;&lt;/div&gt;
   &lt;div id="log"&gt;&lt;/div&gt;
   <span class="highlight">&lt;script&gt;</span>
   <span class="highlight">&lt;/script&gt;</span>
&lt;body&gt;
...
             		        </code>
             		    </pre>
                      </section>

                      <section>
                          <h2>CHECKPOINT</h2>
                          <p class="font2">Starter file should now appear as follows</p>
                          <pre>
                              <code class="html">
&lt;!DOCTYPE html&gt;
&lt;html&gt;
&lt;head&gt;
  &lt;title&gt;CSS Transforms API Test: transform translate&lt;/title&gt;
  &lt;link rel="author" title="Your Name" href="mailto:youremail@address.com"&gt;
  &lt;link rel="help" href="http://www.w3.org/TR/css3-transforms/#transform-property"&gt;
  &lt;link rel="help" href="http://www.w3.org/TR/css3-transforms/#two-d-transform-functions"&gt;
  &lt;meta name="flags" content="dom"&gt;
  &lt;meta name="assert" content="The transform should be translate(100px,100px)"&gt; 
  &lt;script src="/resources/testharness.js"&gt;&lt;/script&gt;
  &lt;script src="/resources/testharnessreport.js"&gt;&lt;/script&gt;    
&lt;/head&gt;
&lt;body&gt;
  &lt;div id="myDiv"&gt;&lt;/div&gt;
  &lt;div id="log"&gt;&lt;/div&gt;
  &lt;script&gt; 
  &lt;/script&gt;    
&lt;/body&gt;
&lt;/html&gt;                              
                              </code>
                          </pre>
                      </section>

                      <section data-state="screenshotPage">
                           <h2>Step 7</h2>
                           <br>
                           <p class="font2">Set the transform on the test element in JavaScript</p>
                           <br>
                           <pre style="width:800px">
                               <code class="html">
...
&lt;body&gt;
   &lt;div id="myDiv"&gt;&lt;/div&gt;
   &lt;div id="log"&gt;&lt;/div&gt;
   &lt;script&gt;
       // Set the transform 
       <span class="highlight">document.getElementById("myDiv").style.transform = "translate(100px,100px)";</span> 
   &lt;/script&gt;    
&lt;/body&gt;
&lt;/html&gt;
                            </code>
                        </pre>
                      </section>

                      <section data-state="screenshotPage">
          			     <h2>Step 8</h2>
              		     <p class="font2">Add a call to test() to your script block to create a test case</p>
              		     <p class="font3">Use assert_equals as the test_function</p>
                         <pre style="width:1000px">
                             <code class="html">
...
&lt;body&gt;
&lt;div id="myDiv"&gt;&lt;/div&gt;
&lt;div id="log"&gt;&lt;/div&gt;
&lt;script&gt;
// Set the transform 
document.getElementById("myDiv").style.transform = "translate(100px,100px)";

// Verify that the transform was set as expected
test(function() {assert_equals( //test arg1: test_function
        document.getElementById("myDiv").style.getPropertyCSSValue("transform").cssText, //assert_equals arg 1: actual
        "translate(100px,100px)", //assert_equals arg 2: expected
        "transform should be translate(100px,100px)")}, //assert_equals arg 3: description
        "Transform_Translate_100px_100px");  //test arg2: name    
&lt;/script&gt;    
&lt;/body&gt;
&lt;/html&gt;
                            </code>
                         </pre>
                        <p class="font3">Comments on test arguments added for clarity, can be omitted</p>  
          			</section>    	

          			<section data-state="screenshotPage">
                           <h2>Completed File</h2>
                           <pre>
                               <code class="html" style="font-size:14px">
&lt;!DOCTYPE html&gt;
&lt;html&gt;
&lt;head&gt;
   &lt;title&gt;CSS Transforms API Test: transform translate&lt;/title&gt;
   &lt;link rel="author" title="Your Name" href="mailto:youremail@address.com"&gt;
   &lt;link rel="help" href="http://www.w3.org/TR/css3-transforms/#transform-property"&gt;
   &lt;link rel="help" href="http://www.w3.org/TR/css3-transforms/#two-d-transform-functions"&gt;
   &lt;meta name="flags" content="dom"&gt;
   &lt;meta name="assert" content="The transform should be translate(100px,100px)"&gt; 
   &lt;script src="/resources/testharness.js"&gt;&lt;/script&gt;
   &lt;script src="/resources/testharnessreport.js"&gt;&lt;/script&gt;    
&lt;/head&gt;
&lt;body&gt;
   &lt;div id="myDiv"&gt;&lt;/div&gt;
   &lt;div id="log"&gt;&lt;/div&gt;
   &lt;script&gt;
       document.getElementById("myDiv").style.transform = "translate(100px,100px)";
       test(function() {assert_equals(
               document.getElementById("myDiv").style.getPropertyCSSValue("transform").cssText,
               "translate(100px,100px)", "transform should be translate(100px,100px)")},
               "Transform_Translate_100px_100px");
   &lt;/script&gt;    
&lt;/body&gt;
&lt;/html&gt;                                   
                               </code>
                           </pre>
                       </section>

                       <section data-state="screenshotPage">
                   	     <h2>Passing test example</p>
                   	     <img src="assets/images/jsTest_PASS.png">
                       </section>

                       <section data-state="screenshotPage">
                   	     <h2>Failing test example</h2>
                         <img src="assets/images/jsTest_FAIL.png">
                       </section>

                       <section data-state="screenshotPage">
                           <section>
                   	           <h2>Step 10 - OSX</h2>
                   	           <p class="font5">Press the down-arrow for Windows instructions</p>
                   		       <p class="font3">Add your new test file to the repository</p>
                   		       <pre>
                   		           <code style="font-size:18px; line-height:22px; height:120px">
$ cd <span class="localPath">[path to repo]</span>/test/contributors/ttwf/<span class="localPath">yourName</span>/incoming
$ hg add ttwf-js-tutorial-001.html
$ hg commit -m "TTWF JavaScript tutorial"
$ hg push
                   		            </code>
                   		       </pre>
           		               <p class="font5" style="line-height:24px">If you receive the error:
           		               <span class="errorText">abort: push creates new remote head ...</span><br>
           		            Your local repository is out of sync with the test repository. To resolve:<br></p>
                   		        <pre>
                   		        <code style="font-size:18px; line-height:22px; height:120px">
$ hg pull
$ hg merge
$ hg commit -m "Merge"
$ hg push
                   		        </code>
                   		    </pre>         			       
                   	    </section>
                   	    <section>
                   	        <h2>Step 10 - Windows</h2>
                   	        <p class="font5">Press the up-arrow for OSX instructions</p>
                   		    <p class="font3">Add your new test file to the repository</p>
               		       <pre>
               		           <code style="font-size:18px; line-height:22px; height:120px">
> cd <span class="localPath">[path to repo]</span>\test\contributors\ttwf\<span class="localPath">yourName</span>\incoming
> hg add ttwf-js-tutorial-001.html
> hg commit -m "TTWF JavaScript tutorial"
> hg push
               		            </code>
               		       </pre>
       		               <p class="font5" style="line-height:24px">If you receive the error:
       		               <span class="errorText">abort: push creates new remote head ...</span><br>
       		            Your local repository is out of sync with the test repository. To resolve:<br></p>
               		        <pre>
               		        <code style="font-size:18px; line-height:22px; height:120px">
> hg pull
> hg merge
> hg commit -m "Merge"
> hg push
               		        </code>
               		    </pre>     			               			    
                   	    </section>
                       </section>

                       <section data-state="screenshotPage">
                   	     <h2>Step 11</h2>
                   	     <br><br><br>
                   	     <h3>Ask the spec expert to review your test</h3>
                       </section>
                       
                       <section data-state="screenshotPage">
                           <section>
                           	    <h2>Step 12 - OSX</h2>
                           	    <p class="font5">Press down-arrow for Windows instructions</p>
                           	    <p class="font3">Incorporate feedback & move test to the submitted folder</p>
                           	    <br>
                           	    <pre>
                           	        <code style="font-size:18px; line-height:22px; height:120px">
$ cd <span class="localPath">[path to repo]</span>/test/contributors/ttwf/<span class="localPath">yourName</span>
$ hg mv incoming/ttwf-js-tutorial-001.html submitted/ttwf-js-tutorial-001.html
$ hg commit -m "moved the TTWF JavaScript tutorial to the submitted folder"
$ hg push                       	            
                           	        </code>
                           	    </pre>
                           </section>
                           
                           <section>
                               <h2>Step 12 - Windows</h2>
                       	       <p class="font5">Press down-arrow for OSX instructions</p>
                           	   <p class="font2">Incorporate feedback & move test to the submitted folder</p>
                          	   <pre>
                      	           <code style="font-size:18px; line-height:22px; height:120px">
> cd <span class="localPath">[path to repo]</span>\test\contributors\ttwf\<span class="localPath">yourName</span>
> hg mv incoming\ttwf-js-tutorial-001.html submitted\ttwf-js-tutorial-001.html
> hg commit -m "moved the TTWF JavaScript tutorial to the submitted folder"
> hg push                       	            
                      	            </code>
                      	        </pre>
                           </section>          		        
                       </section>  

                       <section data-state="screenshotPage">
                           <br><br>
                           <h2>Congratulations!</h2>
                           <br><br>
                           <p class="font1">You have completed both a W3C RefTest and JavaScript Test</p>
                       </section>                 		    		     

                       <section>
                           <br><br><br><br><br>
                           <h2>Reference Slides</h2>
                       </section>

                       <section data-state="screenshotPage">
                           <h2>Mercurial Command Reference</h2>
                           <br>
                           <section>
                               <ul>
                                   <li>Basic Commands:
                                       <ul>
                                           <li>hg pull   (gets the latest changes from the CSSWG Test repo)</li>
                                           <ul>
                                               <li>To update after pull, use: hg pull -u (alleviates need to do hg update)</li>
                                           </ul>                                
                                           <li>hg update (applies the latest changes pulled to your local repo)</li>
                                           <li>hg status (displays list of locally changed files)</li>
                                           <li>hg add    (stages a new or modified local file for commit)</li>
                                           <li>hg remove (stages the removal of a local file for commit)</li>
                                           <li>hg merge  (merges local changes with updates pulled from CSSWG Test repo)</li>
                                           <li>hg commit (commits local changes to local repository)</li>
                                               <ul>
                                                   <li>To include commit message, use: hg commit -m "<span style="font-style:italic">Commit message here</span>"</li>
                                               </ul>
                                           <li>hg push   (pushes local changes to the CSSWG Test repository)</li>
                                       </ul>
                                   </li>
                                </ul>
                                <br><br>
                                <p class="font5">Press down for more</p>
                           </section>
                           <section>
                               <ul>
                                   <li>User wants to update their local repo:
                                       <ul>
                                           <li>hg pull -u  (pulls and applies latest changes from CSSWG repo to local repo)</li>
                                       </ul>
                                   </li>
                                   <li>User wants to push local changes:
                                       <ul>
                                           <li>hg status (check which files are stages for commit)</li>
                                           <li>hg add <span style="font-style:italic">fileName</span> (stages file for commit, repeat for each file)</li>
                                           <li>hg status (confirm all files are stages for commit)</li>
                                           <li>hg commit -m "<span style="font-style:italic">Commit message here</span>" (Commit to local repo)</li>
                                           <li>hg push (pushes locally committed changes to CSSWG Test repo)
                                       </ul>
                                   </li> 
                                </ul>
                                <br><br><br>
                                <p class="font5">Press down for more</p>
                            </section>
                            <section>
                               <ul>
                                   <li>User gets error message during push
                                       <pre>
                                           <code style="font-size:18px">
abort: push creates new remote head ...
                                            </code>
                                        </pre>
                                        <ul>
                                            <li>hg pull (gets newest changes from CSSWG Test repo)</li>
                                            <li>hg merge (merges newest changes from CSSWG Test repo with local changes)</li>
                                            <li>hg commit -m "Merge" (commits the merge to local repo)</li>
                                            <li>hg push (pushes changes to CSSWG Test repo)</li>
                                        </ul>
                                    </li>
                                </ul>
                            </section>
                        </section>
			</div>

			<!-- The navigational controls UI -->
			<aside class="controls">
				<a class="left" href="#">&#x25C4;</a>
				<a class="right" href="#">&#x25BA;</a>
				<a class="up" href="#">&#x25B2;</a>
				<a class="down" href="#">&#x25BC;</a>
			</aside>

			<!-- Displays presentation progress, max value changes via JS to reflect # of slides -->
			<div class="progress"><span></span></div>
			
		</div>

		<!-- Optional libraries for code syntax highlighting and classList support in IE9 -->
		<script src="lib/highlight.js"></script>
		<script src="lib/classList.js"></script>
		
		<script src="js/reveal.js"></script>
		
		<script>
			// Parse the query string into a key/value object
			var query = {};
			location.search.replace( /[A-Z0-9]+?=(\w*)/gi, function(a) {
				query[ a.split( '=' ).shift() ] = a.split( '=' ).pop();
			} );

			// Fires when a slide with data-state=customevent is activated
			Reveal.addEventListener( 'customevent', function() {
				alert( '"customevent" has fired' );
			} );

			// Fires each time a new slide is activated
			Reveal.addEventListener( 'slidechanged', function( event ) {
				// event.indexh & event.indexv
			} );

			Reveal.initialize({
				// Display controls in the bottom right corner
				controls: true,

				// Display a presentation progress bar
				progress: true,

				// If true; each slide will be pushed to the browser history
				history: true,

				// Loops the presentation, defaults to false
				loop: false,

				// Flags if mouse wheel navigation should be enabled
				mouseWheel: true,

				// Apply a 3D roll to links on hover
				rollingLinks: true,

				// UI style
				theme: query.theme || 'default', // default/neon

				// Transition style
				transition: query.transition || 'default' // default/cube/page/concave/linear(2d)
			});

			hljs.initHighlightingOnLoad();
		</script>

	</body>
</html>
